+++ x/ibc/20-transfer/module.go
@@ -1,0 +1,18 @@
+package transfer
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/x/ibc/20-transfer/client/cli"
+)
+
+// Name returns the IBC connection ICS name
+func Name() string {
+	return SubModuleName
+}
+
+// GetTxCmd returns the root tx command for the IBC connections.
+func GetTxCmd(cdc *codec.Codec) *cobra.Command {
+	return cli.GetTxCmd(cdc)
+}
