+++ x/ibc/20-transfer/types/expected_keepers.go
@@ -1,0 +1,25 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	channel "github.com/cosmos/cosmos-sdk/x/ibc/04-channel"
+	"github.com/cosmos/cosmos-sdk/x/ibc/04-channel/exported"
+)
+
+// BankKeeper expected IBC bank keeper
+type BankKeeper interface {
+	AddCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, sdk.Error)
+
+	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) sdk.Error
+
+	SubtractCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, sdk.Error)
+}
+
+// ChannelKeeper expected IBC channel keeper
+type ChannelKeeper interface {
+	GetChannel(ctx sdk.Context, srcPort, srcChan string) (channel channel.Channel, found bool)
+
+	GetNextSequenceSend(ctx sdk.Context, portID, channelID string) (uint64, bool)
+
+	SendPacket(ctx sdk.Context, packet exported.PacketI) error
+}
