+++ x/ibc/20-transfer/types/keys.go
@@ -1,0 +1,15 @@
+package types
+
+const (
+	// SubModuleName defines the IBC transfer name
+	SubModuleName = "transfer"
+
+	// StoreKey is the store key string for IBC transfer
+	StoreKey = SubModuleName
+
+	// RouterKey is the message route for IBC transfer
+	RouterKey = SubModuleName
+
+	// QuerierRoute is the querier route for IBC transfer
+	QuerierRoute = SubModuleName
+)
