+++ x/ibc/mock/bank/client/cli/tx.go
@@ -13,10 +13,10 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/client/utils"
+	channel "github.com/cosmos/cosmos-sdk/x/ibc/04-channel"
 	ics23 "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment"
 	"github.com/cosmos/cosmos-sdk/x/ibc/mock/bank/internal/types"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 func GetTxCmd(cdc *codec.Codec) *cobra.Command {
@@ -26,53 +26,10 @@ func GetTxCmd(cdc *codec.Codec) *cobra.Command {
 		// RunE:  client.ValidateCmd,
 	}
 	txCmd.AddCommand(
-		GetTransferTxCmd(cdc),
 		GetMsgRecvPacketCmd(cdc),
 	)
 
 	return txCmd
-}
-
-func GetTransferTxCmd(cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "transfer --src-port <src port> --src-channel <src channel> --denom <denomination> --amount <amount> --receiver <receiver> --source <source>",
-		Short: "Transfer tokens across chains through IBC",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			txBldr := auth.NewTxBuilderFromCLI().WithTxEncoder(utils.GetTxEncoder(cdc))
-			ctx := context.NewCLIContext().WithCodec(cdc).WithBroadcastMode(flags.BroadcastBlock)
-
-			sender := ctx.GetFromAddress()
-			receiver := viper.GetString(FlagReceiver)
-			denom := viper.GetString(FlagDenom)
-			srcPort := viper.GetString(FlagSrcPort)
-			srcChan := viper.GetString(FlagSrcChannel)
-			source := viper.GetBool(FlagSource)
-
-			amount, ok := sdk.NewIntFromString(viper.GetString(FlagAmount))
-			if !ok {
-				return fmt.Errorf("invalid amount")
-			}
-
-			msg := types.NewMsgTransfer(srcPort, srcChan, denom, amount, sender, receiver, source)
-			if err := msg.ValidateBasic(); err != nil {
-				return err
-			}
-
-			return utils.GenerateOrBroadcastMsgs(ctx, txBldr, []sdk.Msg{msg})
-		},
-	}
-
-	cmd.Flags().AddFlagSet(FsTransfer)
-
-	cmd.MarkFlagRequired(FlagSrcPort)
-	cmd.MarkFlagRequired(FlagSrcChannel)
-	cmd.MarkFlagRequired(FlagDenom)
-	cmd.MarkFlagRequired(FlagAmount)
-	cmd.MarkFlagRequired(FlagReceiver)
-
-	cmd = client.PostCommands(cmd)[0]
-
-	return cmd
 }
 
 // GetMsgRecvPacketCmd returns the command to create a MsgRecvTransferPacket transaction
@@ -85,14 +42,15 @@ func GetMsgRecvPacketCmd(cdc *codec.Codec) *cobra.Command {
 			txBldr := auth.NewTxBuilderFromCLI().WithTxEncoder(utils.GetTxEncoder(cdc))
 			cliCtx := context.NewCLIContext().WithCodec(cdc).WithBroadcastMode(flags.BroadcastBlock)
 
-			var packet types.Packet
+			var packet channel.Packet
 			if err := cdc.UnmarshalJSON([]byte(args[0]), &packet); err != nil {
 				fmt.Fprintf(os.Stderr, "failed to unmarshall input into struct, checking for file...\n")
 				contents, err := ioutil.ReadFile(args[0])
 				if err != nil {
 					return fmt.Errorf("error opening packet file: %v", err)
 				}
-				if err := packet.UnmarshalJSON(contents); err != nil {
+
+				if err := cdc.UnmarshalJSON(contents, packet); err != nil {
 					return fmt.Errorf("error unmarshalling packet file: %v", err)
 				}
 			}
