+++ x/ibc/mock/bank/internal/types/errors.go
@@ -6,12 +6,11 @@ import (
 
 // ibcmockbank errors reserve 100 ~ 199.
 const (
-	CodeIncorrectDenom    sdk.CodeType = 101
-	CodeInvalidAmount     sdk.CodeType = 102
-	CodeInvalidAddress    sdk.CodeType = 103
-	CodeInvalidReceiver   sdk.CodeType = 104
-	CodeErrSendPacket     sdk.CodeType = 105
-	CodeErrReceivePacket  sdk.CodeType = 106
-	CodeProofMissing      sdk.CodeType = 107
-	CodeInvalidPacketData sdk.CodeType = 108
+	DefaultCodespace sdk.CodespaceType = ModuleName
+
+	CodeInvalidAmount     sdk.CodeType = 101
+	CodeInvalidAddress    sdk.CodeType = 102
+	CodeErrReceivePacket  sdk.CodeType = 103
+	CodeProofMissing      sdk.CodeType = 104
+	CodeInvalidPacketData sdk.CodeType = 105
 )
