+++ x/ibc/mock/bank/internal/types/keys.go
@@ -7,15 +7,9 @@ const (
 	// StoreKey is the string store representation
 	StoreKey = ModuleName
 
-	// TStoreKey is the string transient store representation
-	TStoreKey = "transient_" + ModuleName
-
 	// QuerierRoute is the querier route for the module
 	QuerierRoute = ModuleName
 
 	// RouterKey is the msg router key for the module
 	RouterKey = ModuleName
-
-	// codespace
-	DefaultCodespace = ModuleName
 )
